fix(交易): 修复玩家点数类型转换并添加经济管理器依赖
在TradeEconomyManager中修复getPoints方法的返回值类型转换问题，将double转为int。同时修改TradeCommand构造函数和TradeGUI初始化，添加TradeEconomyManager依赖以支持经济功能。

+++ server/src/main/java/com/oolonghoo/woosocial/module/trade/TradeEconomyManager.java
@@ -66,7 +66,7 @@ public class TradeEconomyManager {
     
     public int getPoints(Player player) {
         if (!hasPlayerPoints()) return 0;
-        return playerPointsHook.getBalance(player);
+        return (int) playerPointsHook.getBalance(player);
     }
     
     public boolean withdrawMoney(Player player, double amount) {
